requirements.txt
lazyllm==0.4.1
beautifulsoup4
python-dotenv